+++ src/front/js/pages/pages_usuarios/edicionuser.js
@@ -1,36 +1,32 @@
 import React, { useState, useEffect, useCallback } from "react";
 import { useParams, Link } from "react-router-dom";
-import "../../styles/app.scss";
-//Importamos la libreria axios
+import "../../../styles/app.scss";
+//Importamos axios
 import axios from "axios";
-//Componentes importados
-import { Usuario } from "../component/usuario.js";
+
 //Aqui colocar la URL de la API por favor
-import { URL } from "../config";
+import { URL } from "../../config";
+
+//Componentes importados
+import { EditaruserCompont } from "../../component/edicionusercomponent.js";
 
 //react-bootstrap
 import { Container, Form, Button } from "react-bootstrap";
 import "bootstrap/dist/css/bootstrap.min.css";
 
-export const Datosuser = () => {
+export const Editaruser = props => {
 	const [user, setUser] = useState(null);
 	const { id } = useParams();
-
-	/**
-	 * Función para traernos la información del usuario mediante axios
-	 * Después de obtenerlo lo guardamos con la función setUser
-	 * que nos brinda useState
-	 */
 
 	const fetchUser = useCallback(
 		async () => {
 			try {
 				const { data } = await axios.get(`${URL}lista_users/${id}`);
-				setUser(data);
 				// console.log("user", data);
+				setUser(data);
 			} catch (error) {
 				console.error(error);
-				alert("Error en la api: No se pudo recibir la lista de usuarios");
+				alert("Error en la api: No pude traer los datos");
 			}
 		},
 		[setUser]
@@ -38,7 +34,6 @@ export const Datosuser = () => {
 
 	useEffect(
 		() => {
-			// console.log("id", id);
 			fetchUser();
 		},
 		[fetchUser]
@@ -48,8 +43,7 @@ export const Datosuser = () => {
 		<>
 			{user ? (
 				<>
-					<Usuario
-						id={Number(id)}
+					<EditaruserCompont
 						perfil={user.perfil}
 						nombre={user.name}
 						apellido={user.lastname}
@@ -62,15 +56,13 @@ export const Datosuser = () => {
 					<Container>
 						<Form>
 							<Button className="my-3" variant="primary">
-								<Link className="text-light" to={`../editar_user/${Number(id)}`}>
-									Editar
-								</Link>
+								<Link className="text-light">Enviar</Link>
 							</Button>{" "}
 						</Form>
 					</Container>
 				</>
 			) : (
-				<h2>Cargando...</h2>
+				<h2>Cargando</h2>
 			)}
 		</>
 	);
